backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and authentication
supabase==2.0.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Microsoft Graph API for OneDrive
msal==1.25.0
requests==2.31.0

# CORS middleware
python-multipart==0.0.6

# AI and document processing
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
beautifulsoup4==4.12.2
xlrd==2.0.1
lxml==4.9.3

# Utilities
pydantic==2.5.0
python-dotenv==1.0.0